+++ internal/crawler/core/http_client.go
@@ -39,15 +39,22 @@ type StandardHTTPClient struct {
 	client      *http.Client
 	userAgent   string
 	timeout     time.Duration
-	rateLimiter *IPRateLimiterRegistry
+	rateLimiter URLRateLimiter
 }
 
 // NewHTTPClient는 새로운 HTTP 클라이언트를 생성합니다.
-// Config의 RequestsPerHour, BurstSize로 IP 기반 rate limiter를 초기화합니다.
+// rate limiter 없이 동작하며, WithRateLimiter로 주입할 수 있습니다.
 func NewHTTPClient(config Config) HTTPClient {
-	resolver := NewDNSIPResolver(5 * time.Minute)
-	rateLimiter := NewIPRateLimiterRegistry(resolver, config.RequestsPerHour, config.BurstSize)
+	return &StandardHTTPClient{
+		client:    defaultHTTPClient(config),
+		userAgent: config.UserAgent,
+		timeout:   config.Timeout,
+	}
+}
 
+// NewHTTPClientWithRateLimiter는 URLRateLimiter가 주입된 HTTP 클라이언트를 생성합니다.
+// HTTP 요청 전 목적지 기반 rate limiting을 적용합니다.
+func NewHTTPClientWithRateLimiter(config Config, rateLimiter URLRateLimiter) HTTPClient {
 	return &StandardHTTPClient{
 		client:      defaultHTTPClient(config),
 		userAgent:   config.UserAgent,
@@ -60,9 +67,11 @@ func NewHTTPClient(config Config) HTTPClient {
 func (c *StandardHTTPClient) Get(ctx context.Context, url string) (*HTTPResponse, error) {
 	log := logger.FromContext(ctx)
 
-	// IP 기반 rate limiting: 요청 전 목적지 IP의 rate limiter에서 대기
-	if err := c.rateLimiter.Wait(ctx, url); err != nil {
-		return nil, fmt.Errorf("rate limit wait for %s: %w", url, err)
+	// 목적지 기반 rate limiting: rate limiter가 주입된 경우에만 적용
+	if c.rateLimiter != nil {
+		if err := c.rateLimiter.Wait(ctx, url); err != nil {
+			return nil, fmt.Errorf("rate limit wait for %s: %w", url, err)
+		}
 	}
 
 	start := time.Now()
@@ -107,9 +116,11 @@ func (c *StandardHTTPClient) Get(ctx context.Context, url string) (*HTTPResponse
 func (c *StandardHTTPClient) Post(ctx context.Context, url string, body []byte) (*HTTPResponse, error) {
 	log := logger.FromContext(ctx)
 
-	// IP 기반 rate limiting: 요청 전 목적지 IP의 rate limiter에서 대기
-	if err := c.rateLimiter.Wait(ctx, url); err != nil {
-		return nil, fmt.Errorf("rate limit wait for %s: %w", url, err)
+	// 목적지 기반 rate limiting: rate limiter가 주입된 경우에만 적용
+	if c.rateLimiter != nil {
+		if err := c.rateLimiter.Wait(ctx, url); err != nil {
+			return nil, fmt.Errorf("rate limit wait for %s: %w", url, err)
+		}
 	}
 
 	start := time.Now()
