+++ test/internal/rate_limiter/dns_resolver_test.go
@@ -1,4 +1,4 @@
-package core_test
+package rate_limiter_test
 
 import (
 	"testing"
@@ -7,25 +7,20 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	core "issuetracker/internal/crawler/core"
+	ratelimiter "issuetracker/internal/crawler/rate_limiter"
 )
 
-// ─────────────────────────────────────────────────────────────────────────────
-// DNSIPResolver 테스트
-// ─────────────────────────────────────────────────────────────────────────────
-
 func TestDNSIPResolver_Resolve_Localhost(t *testing.T) {
-	resolver := core.NewDNSIPResolver(5 * time.Minute)
+	resolver := ratelimiter.NewDNSIPResolver(5 * time.Minute)
 
 	ip, err := resolver.Resolve("http://localhost:8080/path")
 	require.NoError(t, err)
 	assert.NotEmpty(t, ip)
-	// localhost는 127.0.0.1 또는 ::1로 해석됨
 	assert.Contains(t, []string{"127.0.0.1", "::1"}, ip)
 }
 
 func TestDNSIPResolver_Resolve_CachesResult(t *testing.T) {
-	resolver := core.NewDNSIPResolver(1 * time.Hour)
+	resolver := ratelimiter.NewDNSIPResolver(1 * time.Hour)
 
 	ip1, err := resolver.Resolve("http://localhost/a")
 	require.NoError(t, err)
@@ -38,21 +33,21 @@ func TestDNSIPResolver_Resolve_CachesResult(t *testing.T) {
 }
 
 func TestDNSIPResolver_Resolve_InvalidURL_ReturnsError(t *testing.T) {
-	resolver := core.NewDNSIPResolver(5 * time.Minute)
+	resolver := ratelimiter.NewDNSIPResolver(5 * time.Minute)
 
 	_, err := resolver.Resolve("://invalid")
 	assert.Error(t, err)
 }
 
 func TestDNSIPResolver_Resolve_EmptyHost_ReturnsError(t *testing.T) {
-	resolver := core.NewDNSIPResolver(5 * time.Minute)
+	resolver := ratelimiter.NewDNSIPResolver(5 * time.Minute)
 
 	_, err := resolver.Resolve("http:///path-only")
 	assert.Error(t, err)
 }
 
 func TestDNSIPResolver_Resolve_UnresolvableHost_ReturnsError(t *testing.T) {
-	resolver := core.NewDNSIPResolver(5 * time.Minute)
+	resolver := ratelimiter.NewDNSIPResolver(5 * time.Minute)
 
 	_, err := resolver.Resolve("http://this-host-does-not-exist-12345.invalid/path")
 	assert.Error(t, err)
